Add files via upload
The file  and code runs and works in other open project files, but still gives a unfixable module error where it says matplotlib's 'module' cannot be identified. Running it using the setup in the daily work causes the error to occur but it can be worked around by transferring it to another open file in Pycharm. I've tried re-installing things and working around it but it cannot be permanently fixed.